Conflicts:
	Final Project/FrontEnd.X/SchedMain.c
	Final Project/FrontEnd.X/game.c
	Final Project/FrontEnd.X/nbproject/Makefile-genesis.properties
Took theirs

+++ Final Project/FrontEnd.X/globals.h
@@ -38,6 +38,7 @@ extern "C" {
 #include "motorDriver.h"
 #include "game.h"
 #include "i2cComm.h"
+#include "xbee.h"
 
 #define FRONT_NOT_BACK 1 // is this the Front end or backend (should move to globals)
 #define CARDSLOT_1 0b00010000
@@ -66,7 +67,7 @@ extern "C" {
         int mode;
 	int cardSelect[4];
 	int selectMove[4][3];
-        short game;
+        int game;
         Boolean getInventory;
         Boolean sendI2C; // i2c command prepped for transmission
         Boolean gotI2C; // i2c command was received
